fix(sentrux): repin sentrux_gate.rs, add end-to-end boundary/layer tests
CI caught two real gaps in the previous commit:

- orchestration/internalization/sentrux.json still pinned
  sentrux_gate.rs to its pre-edit sha256; `code-intel repin --write`
  resyncs it.
- boundary_rules.rs's tests lived inline instead of in a sibling
  `_tests.rs` file, so this crate's own is_test_file heuristic didn't
  credit them -- change-risk scored the PR "no tests touched" despite
  the coverage existing. Split to boundary_rules_tests.rs (this
  crate's established convention) and add two fixture-driven
  `run_check` tests that create a real two-file tree with an actual
  `use crate::cli` edge: one asserting the forbidden edge is caught,
  one asserting the allowed direction isn't. The prior self-check test
  only proved this repository has no violation today -- it would stay
  green even if `crate_edges` regressed to returning nothing.
