+++ crates/code-intel-cli/src/boundary_rules.rs
@@ -261,156 +261,5 @@ fn array_field(table: &BTreeMap<String, String>, key: &str) -> Vec<String> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::path::Path;
-
-    #[test]
-    fn table_array_reads_one_table_per_marker_and_stops_at_the_next_header() {
-        let rules = r#"
-[constraints]
-max_cc = 156
-
-[[boundary]]
-description = "engine must not depend on cli"
-from = ["sentrux_gate", "sentrux_analysis"]
-forbid = ["cli"]
-
-[[boundary]]
-description = "second rule"
-from = ["a"]
-forbid = ["b", "c"]
-
-[[layer]]
-modules = ["sentrux_gate"]
-"#;
-        let boundaries = parse_boundaries(rules);
-        assert_eq!(boundaries.len(), 2);
-        assert_eq!(boundaries[0].description, "engine must not depend on cli");
-        assert_eq!(boundaries[0].from, vec!["sentrux_gate", "sentrux_analysis"]);
-        assert_eq!(boundaries[0].forbid, vec!["cli"]);
-        assert_eq!(boundaries[1].forbid, vec!["b", "c"]);
-
-        let layers = parse_layers(rules);
-        assert_eq!(layers.len(), 1);
-        assert_eq!(layers[0].modules, vec!["sentrux_gate"]);
-    }
-
-    #[test]
-    fn owning_module_resolves_nested_files_to_their_top_level_module() {
-        let rust_files: BTreeSet<String> = [
-            "crates/x/src/cli/mod.rs",
-            "crates/x/src/cli/legacy.rs",
-            "crates/x/src/sentrux_gate.rs",
-        ]
-        .into_iter()
-        .map(String::from)
-        .collect();
-        assert_eq!(
-            owning_module("crates/x/src/cli/legacy.rs", &rust_files),
-            Some("cli".to_string())
-        );
-        assert_eq!(
-            owning_module("crates/x/src/sentrux_gate.rs", &rust_files),
-            Some("sentrux_gate".to_string())
-        );
-        // A nested file whose first path segment names neither `<segment>.rs`
-        // nor `<segment>/mod.rs` in the tree resolves to no real top-level
-        // crate module, so it is left unclassified rather than guessed.
-        assert_eq!(
-            owning_module("crates/x/src/ghost/deep.rs", &rust_files),
-            None
-        );
-    }
-
-    #[test]
-    fn boundary_violations_only_fire_for_the_declared_direction() {
-        let mut edges: BTreeMap<String, BTreeSet<String>> = BTreeMap::new();
-        edges
-            .entry("crates/x/src/sentrux_gate.rs".into())
-            .or_default()
-            .insert("crates/x/src/cli/mod.rs".into());
-        let mut modules_by_file = BTreeMap::new();
-        modules_by_file.insert(
-            "crates/x/src/sentrux_gate.rs".to_string(),
-            "sentrux_gate".to_string(),
-        );
-        modules_by_file.insert("crates/x/src/cli/mod.rs".to_string(), "cli".to_string());
-
-        let rules = vec![BoundaryRule {
-            description: "core must not depend on cli".into(),
-            from: vec!["sentrux_gate".into()],
-            forbid: vec!["cli".into()],
-        }];
-        let violations = boundary_violations(&rules, &edges, &modules_by_file);
-        assert_eq!(violations.len(), 1);
-        assert_eq!(violations[0].rule, "boundary_dependency");
-        assert_eq!(
-            violations[0].targets,
-            vec!["crates/x/src/sentrux_gate.rs -> crates/x/src/cli/mod.rs"]
-        );
-
-        // The reverse direction (cli -> sentrux_gate) is exactly what's
-        // expected of an outer layer and must never trip this rule.
-        let mut reverse_edges: BTreeMap<String, BTreeSet<String>> = BTreeMap::new();
-        reverse_edges
-            .entry("crates/x/src/cli/mod.rs".into())
-            .or_default()
-            .insert("crates/x/src/sentrux_gate.rs".into());
-        assert!(boundary_violations(&rules, &reverse_edges, &modules_by_file).is_empty());
-    }
-
-    #[test]
-    fn layer_violations_only_fire_when_an_earlier_layer_depends_on_a_later_one() {
-        let layers = vec![
-            Layer {
-                modules: vec!["core".into()],
-            },
-            Layer {
-                modules: vec!["cli".into()],
-            },
-        ];
-        let mut modules_by_file = BTreeMap::new();
-        modules_by_file.insert("src/core.rs".to_string(), "core".to_string());
-        modules_by_file.insert("src/cli.rs".to_string(), "cli".to_string());
-
-        let mut forward_edges: BTreeMap<String, BTreeSet<String>> = BTreeMap::new();
-        forward_edges
-            .entry("src/cli.rs".into())
-            .or_default()
-            .insert("src/core.rs".into());
-        assert!(layer_violations(&layers, &forward_edges, &modules_by_file).is_empty());
-
-        let mut backward_edges: BTreeMap<String, BTreeSet<String>> = BTreeMap::new();
-        backward_edges
-            .entry("src/core.rs".into())
-            .or_default()
-            .insert("src/cli.rs".into());
-        let violations = layer_violations(&layers, &backward_edges, &modules_by_file);
-        assert_eq!(violations.len(), 1);
-        assert_eq!(violations[0].rule, "layer_order");
-        assert_eq!(violations[0].targets, vec!["src/core.rs -> src/cli.rs"]);
-    }
-
-    #[test]
-    fn this_repository_declares_no_boundary_or_layer_violation() {
-        // Same self-check convention as
-        // `this_repository_has_no_resolved_import_cycles`: run the real
-        // engine against this crate's own tree, using whatever
-        // `.sentrux/rules.toml` currently declares.
-        let repo = Path::new(env!("CARGO_MANIFEST_DIR")).join("../..");
-        let run = super::super::run_check(&repo)
-            .expect("sentrux check should run against this repository");
-        let violations: Vec<&Violation> = run
-            .violations
-            .iter()
-            .filter(|violation| {
-                violation.rule == "boundary_dependency" || violation.rule == "layer_order"
-            })
-            .collect();
-        assert!(
-            violations.is_empty(),
-            "boundary_dependency/layer_order violation(s) detected: {violations:?}"
-        );
-    }
-}
+#[path = "boundary_rules_tests.rs"]
+mod tests;
