+++ orchestration/internalization/sentrux.json
@@ -258,7 +258,7 @@
       },
       "source": {
         "path": "crates/code-intel-cli/src/sentrux_gate.rs",
-        "sha256": "e4a641630fe293cf7cbc5cb21a464cc0a16f29e3e8491d4c19eb0813b7b9c2b2"
+        "sha256": "f69776e6800f917099794a5a28dbdb1117eaf657568dd09e34bf33e4207ddb4c"
       },
       "conformance": {
         "path": "crates/code-intel-cli/tests/dag_run.rs",
@@ -354,7 +354,7 @@
       "path": "crates/code-intel-cli/src/sentrux_gate.rs",
       "description": "Pipeline-owned built-in structural gate engine (metrics, rules check, no-degradation gate, cycle detection)",
       "evidenceIds": [
-        "local:b03:native-gate-source-sha256:e4a641630fe293cf7cbc5cb21a464cc0a16f29e3e8491d4c19eb0813b7b9c2b2",
+        "local:b03:native-gate-source-sha256:f69776e6800f917099794a5a28dbdb1117eaf657568dd09e34bf33e4207ddb4c",
         "local:b03:native-gate-conformance:ba171006a99900a4a6309353429af397a4b34e37971f8dccf8f0b408cac57f8e"
       ]
     }
